chore: refactor requester package into separate files
- to make the single responsibility principle clearer, the package
  is separated into different files, each with one responsibility
- attributes on the Requester that should be private are made private
- error parsing made tidier using object-orientation
- tests tidied up, and an implementation test (that tested what should be
  internals, rather than behavior) was removed

[#186464017](https://www.pivotaltracker.com/story/show/186464017)
