+++ internal/requester/requester_test.go
@@ -12,7 +12,6 @@ import (
 )
 
 var _ = Describe("Requester", func() {
-
 	var (
 		fakeRequester requester.Requester
 		fakeServer    *ghttp.Server
@@ -28,15 +27,6 @@ var _ = Describe("Requester", func() {
 		fakeRequester = requester.NewRequester(fakeServer.URL(), "fake-token", false)
 	})
 
-	Describe("NewRequester", func() {
-		It("returns a requester with given values", func() {
-			actualRequester := requester.NewRequester("test-url", "fake-token", false)
-
-			Expect(actualRequester.APIBaseURL).To(Equal("test-url"))
-			Expect(actualRequester.APIToken).To(Equal("fake-token"))
-		})
-	})
-
 	Describe("Get", func() {
 		When("request is valid", func() {
 			BeforeEach(func() {
@@ -48,6 +38,7 @@ var _ = Describe("Requester", func() {
 					),
 				)
 			})
+
 			It("succeeds", func() {
 				err := fakeRequester.Get("test-endpoint", &testReceiver)
 
@@ -66,14 +57,10 @@ var _ = Describe("Requester", func() {
 					),
 				)
 			})
+
 			It("returns an error", func() {
 				err := fakeRequester.Get("not-a-real-url", &testReceiver)
 				Expect(err).To(MatchError("http response: 404"))
-			})
-
-			It("errors if receiver is not of type pointer", func() {
-				err := fakeRequester.Get("test-endpoint", testReceiver)
-				Expect(err).To(MatchError("receiver must be of type Pointer"))
 			})
 		})
 
@@ -87,9 +74,17 @@ var _ = Describe("Requester", func() {
 					),
 				)
 			})
+
 			It("returns an error", func() {
 				err := fakeRequester.Get("test-endpoint", &testReceiver)
 				Expect(err).To(MatchError("failed to unmarshal response into receiver error: unexpected end of JSON input"))
+			})
+		})
+
+		When("passed a receiver that is not a pointer", func() {
+			It("returns an error", func() {
+				err := fakeRequester.Get("test-endpoint", testReceiver)
+				Expect(err).To(MatchError("receiver must be of type Pointer"))
 			})
 		})
 	})
@@ -105,9 +100,11 @@ var _ = Describe("Requester", func() {
 					),
 				)
 			})
+
 			It("succeeds", func() {
 				err := fakeRequester.Patch("test-endpoint", `data`)
 				Expect(err).NotTo(HaveOccurred())
+				Expect(fakeServer.ReceivedRequests()).To(HaveLen(1))
 			})
 		})
 
@@ -122,11 +119,13 @@ var _ = Describe("Requester", func() {
 						),
 					)
 				})
+
 				It("returns an error", func() {
 					err := fakeRequester.Patch("test-endpoint", `data`)
 					Expect(err).To(MatchError("http_error: 500 Internal Server Error response_body: Some body"))
 				})
 			})
+
 			When("fails with capi error", func() {
 				BeforeEach(func() {
 					fakeServer.AppendHandlers(
@@ -137,6 +136,7 @@ var _ = Describe("Requester", func() {
 						),
 					)
 				})
+
 				It("returns an error", func() {
 					err := fakeRequester.Patch("test-endpoint", `data`)
 					Expect(err).To(MatchError("http_error: 500 Internal Server Error capi_error_code: 10008 capi_error_title: error title capi_error_detail: error detail"))
@@ -153,6 +153,7 @@ var _ = Describe("Requester", func() {
 						),
 					)
 				})
+
 				It("returns an error", func() {
 					err := fakeRequester.Patch("test-endpoint", `data`)
 					Expect(err.Error()).To(ContainSubstring("http_error: 500 Internal Server Error"))
